Exercise partial run recovery against a real WebDAV failure
