+++ backend/tests/e2e/test_remote_backup.py
@@ -198,86 +198,3 @@ class TestRemoteBackup:
         assert recovered.status_code == 200, recovered.text
         assert recovered.content == FIXTURE.read_bytes()
         assert not list(Path(settings.backup_dir).glob("*.tar.gz"))
-
-
-@pytest.mark.s3
-class TestReplicaRunRecovery:
-    @pytest.mark.critical
-    @pytest.mark.asyncio
-    async def test_partial_run_retries_the_same_archive_through_the_api(
-        self, api, tmp_path
-    ):
-        import hashlib
-
-        from tests.e2e.test_backup import _setup_and_login
-
-        headers = await _setup_and_login(api, tmp_path)
-        endpoint = s3_endpoint()
-        bucket = f"retry-api-{uuid4().hex[:12]}"
-        client = boto3.client(
-            "s3",
-            endpoint_url=endpoint,
-            region_name="us-east-1",
-            aws_access_key_id=S3_ACCESS_KEY,
-            aws_secret_access_key=S3_SECRET_KEY,
-            config=BotoConfig(s3={"addressing_style": "path"}),
-        )
-        connection = await api.post(
-            "/api/v1/storage-connections",
-            headers=headers,
-            json={
-                "name": "Temporarily unavailable replica",
-                "kind": "s3",
-                "purpose": "backup",
-                "configuration": {
-                    "provider": "s3_self_hosted",
-                    "bucket": bucket,
-                    "endpoint_url": endpoint,
-                    "region": "us-east-1",
-                    "root": "replicas",
-                    "addressing_style": "path",
-                },
-                "secrets": {"access_key": S3_ACCESS_KEY, "secret_key": S3_SECRET_KEY},
-            },
-        )
-        assert connection.status_code == 201, connection.text
-        created = await api.post("/api/v1/backups", headers=headers)
-        assert created.status_code == 202, created.text
-        body = created.json()
-        assert body["outcome"] == "partial"
-        failed = next(
-            row for row in body["destination_results"] if row["kind"] == "connection"
-        )
-        local = next(
-            row for row in body["destination_results"] if row["kind"] == "local"
-        )
-        assert local["outcome"] == "completed"
-        assert failed["outcome"] == "failed"
-        client.create_bucket(Bucket=bucket)
-        try:
-            response = await api.post(
-                f"/api/v1/backups/runs/destinations/{failed['id']}/retry",
-                headers=headers,
-            )
-            assert response.status_code == 200, response.text
-            result = response.json()
-            assert result["outcome"] == "completed"
-            assert result["retry_attempts"][0]["source_result_id"] == local["id"]
-            key = result["key"].removeprefix(f"s3/{bucket}/")
-            remote = client.get_object(Bucket=bucket, Key=key)["Body"]
-            try:
-                assert (
-                    hashlib.sha256(remote.read()).hexdigest() == body["archive_sha256"]
-                )
-            finally:
-                remote.close()
-            detail = await api.get(
-                f"/api/v1/backups/runs/{body['run_id']}", headers=headers
-            )
-            assert detail.json()["outcome"] == "completed"
-            assert detail.json()["backup_id"] == body["backup_id"]
-        finally:
-            for page in client.get_paginator("list_objects_v2").paginate(Bucket=bucket):
-                for item in page.get("Contents", []):
-                    client.delete_object(Bucket=bucket, Key=item["Key"])
-            client.delete_bucket(Bucket=bucket)
